Python/VidXXC/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "VidXXC"
description = "Video Encrypt and Decrypt Tool."
version = "0.1.0"
authors = [
  { name = "Jemtaly" },
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "cryptography",
  "Pillow",
  "numpy",
]

[project.optional-dependencies]
cli = [
]
gui = [
  "opencv-python",
]

[project.scripts]
vidxxc-cli = "vidxxc.cli:main [cli]"
vidxxc-gui = "vidxxc.gui:main [gui]"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
